@@ -12,7 +12,7 @@ publish_to: none
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 0.40.6+131
+version: 0.41.0+133
 
 environment:
   sdk: '>=3.0.0 <4.0.0'
@@ -20,7 +20,7 @@ environment:
 dependencies:
   android_id: ^0.3.5
   ansicolor: ^2.0.1
-  archive: ^3.3.8
+  archive: ^3.4.6
   async: ^2.8.1
   audio_session: ^0.1.16
   bloc: ^8.1.2
@@ -44,10 +44,9 @@ dependencies:
       ref: a8e2bab0447760eb09457eafbc49e8d7c8cbb99f
       path: packages/desktop_drop
   desktop_keep_screen_on: ^0.0.2
-  desktop_lifecycle: ^0.1.1
   desktop_webview_window: ^0.2.3
   device_info_plus: ^9.0.3
-  dio: ^5.3.2
+  dio: ^5.3.3
   dio_smart_retry: ^5.0.0
   drift: ^2.11.1
   ecache: ^2.0.2
@@ -65,20 +64,20 @@ dependencies:
   flutter_app_icon_badge: ^2.0.0
   flutter_bloc: ^8.1.3
   flutter_hooks: ^0.20.1
-  flutter_local_notifications: ^15.1.1
+  flutter_local_notifications: ^16.1.0
   flutter_localizations:
     sdk: flutter
   flutter_portal: ^1.1.4
   flutter_svg: ^2.0.7
   gallery_saver: ^2.3.2
-  github: ^9.17.0
+  github: ^9.19.0
   hive: ^2.2.3
   hive_flutter: ^1.0.0
-  hooks_riverpod: ^2.4.0
+  hooks_riverpod: ^2.4.3
   http: ^0.13.6
   http_client_helper: ^2.0.3
   hydrated_bloc: ^9.1.2
-  image: ^4.0.17
+  image: ^4.1.3
   image_picker: ^1.0.4
   intl: ^0.18.0
   intl_phone_number_input: ^0.7.3+1
@@ -91,25 +90,21 @@ dependencies:
       ref: 69bab28cda3bd948f0db5ad7fc9831c29fdcf400
       path: packages/local_auth/local_auth
   lottie: ^2.6.0
-  map: ^1.3.3
-  markdown_widget: ^2.2.0
+  map: ^1.4.2
+  markdown_widget: ^2.3.1
   mime: ^1.0.2
   mixin_bot_sdk_dart: ^0.4.1
-  mixin_logger:
-    git:
-      url: https://github.com/MixinNetwork/flutter-plugins.git
-      ref: 3214026ad280f57b46b738845bc30b404dab26fc
-      path: packages/mixin_logger
+  mixin_logger: ^0.1.0
   native_dio_adapter: ^0.1.0
-  network_info_plus: ^4.0.2
-  octo_image: ^1.0.2
+  network_info_plus: ^4.1.0
+  octo_image: ^2.0.0
   ogg_opus_player: ^0.7.0
   open_file:
     git:
       url: https://github.com/crazecoder/open_file.git
       ref: fbf68e4bb5cb3e262d8f8ebe10b2f8449ff8e030
   overlay_support: ^2.1.0
-  package_info_plus: ^4.1.0
+  package_info_plus: ^4.2.0
   path: ^1.8.0
   path_provider: ^2.1.1
   photo_view:
@@ -118,7 +113,7 @@ dependencies:
       ref: 52685ab7beaf8a9107904bdb8a1590924015d016
   pin_code_fields: ^8.0.1
   platform_device_id: ^1.0.1
-  pretty_qr_code: ^2.0.3
+  pretty_qr_code: ^3.0.0
   protocol_handler: ^0.1.5
   provider: ^6.0.4
   quick_breakpad:
@@ -126,8 +121,9 @@ dependencies:
   recase: ^4.1.0
   rxdart: ^0.27.5
   scrollable_positioned_list: ^0.3.5
+  simple_animations: ^5.0.2
   sliver_tools: ^0.2.12
-  sqlite3_flutter_libs: ^0.5.16
+  sqlite3_flutter_libs: ^0.5.17
   stream_transform: ^2.1.0
   string_tokenizer:
     git:
@@ -152,10 +148,10 @@ dependencies:
   visibility_detector: ^0.4.0+2
   watcher: ^1.1.0
   web_socket_channel: ^2.4.0
-  webview_flutter: ^4.2.4
-  win32: ^5.0.7
+  webview_flutter: ^4.4.1
+  win32: ^5.0.9
   win_toast: ^0.3.0
-  window_manager: ^0.3.6
+  window_manager: ^0.3.7
   window_size:
     git:
       url: https://github.com/google/flutter-desktop-embedding.git
@@ -170,13 +166,13 @@ dev_dependencies:
       url: https://github.com/MixinNetwork/flutter-plugins.git
       ref: 77f03330001941f0ac303de6276dd2380f5bcc7e
       path: packages/debug_info_collector
-  drift_dev: ^2.11.2
+  drift_dev: ^2.12.1
   flutter_test:
     sdk: flutter
   hive_generator: ^2.0.1
   intl_utils: ^2.8.4
   json_serializable: ^6.7.1
-  msix: ^3.16.1
+  msix: ^3.16.4
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
