@@ -1,7 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
-import 'package:flutter_svg/svg.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:mixin_bot_sdk_dart/mixin_bot_sdk_dart.dart';
 import 'package:super_context_menu/super_context_menu.dart';
@@ -132,7 +131,7 @@ class _ParticipantList extends HookConsumerWidget {
   }
 }
 
-class _ParticipantTile extends StatelessWidget {
+class _ParticipantTile extends HookWidget {
   const _ParticipantTile({
     required this.participant,
     required this.currentUser,
@@ -161,19 +160,20 @@ class _ParticipantTile extends StatelessWidget {
           title: Row(
             children: [
               Flexible(
-                child: HighlightText(
+                child: CustomText(
                   participant.fullName ?? '?',
                   style: TextStyle(
                     color: context.theme.text,
                     fontSize: 16,
                   ),
-                  highlightTextSpans: [
-                    HighlightTextSpan(
+                  textMatchers: [
+                    EmojiTextMatcher(),
+                    KeyWordTextMatcher(
                       keyword,
                       style: TextStyle(
                         color: context.theme.accent,
                       ),
-                    )
+                    ),
                   ],
                 ),
               ),
@@ -326,14 +326,14 @@ class _ActionAddParticipants extends HookConsumerWidget {
   final List<ParticipantUser> participants;
 
   @override
-  Widget build(BuildContext context, WidgetRef ref) => PopupMenuPageButton(
+  Widget build(BuildContext context, WidgetRef ref) => CustomPopupMenuButton(
         itemBuilder: (context) => [
-          CustomPopupMenuButton(
+          CustomPopupMenuItem(
             icon: Resources.assetsImagesContextMenuSearchUserSvg,
             title: context.l10n.addParticipants,
             value: _ActionType.addParticipants,
           ),
-          CustomPopupMenuButton(
+          CustomPopupMenuItem(
             icon: Resources.assetsImagesContextMenuLinkSvg,
             title: context.l10n.inviteToGroupViaLink,
             value: _ActionType.inviteByLink,
@@ -372,14 +372,6 @@ class _ActionAddParticipants extends HookConsumerWidget {
               }
           }
         },
-        icon: SvgPicture.asset(
-          Resources.assetsImagesIcAddSvg,
-          height: 24,
-          width: 24,
-          colorFilter: ColorFilter.mode(
-            context.theme.icon,
-            BlendMode.srcIn,
-          ),
-        ),
+        icon: Resources.assetsImagesIcAddSvg,
       );
 }
