@@ -1,4 +1,6 @@
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:intl/intl.dart';
 import 'package:mixin_bot_sdk_dart/mixin_bot_sdk_dart.dart' hide encryptPin;
@@ -214,68 +216,81 @@ class _DeleteAccountPinDialog extends StatelessWidget {
   final String verificationId;
 
   @override
-  Widget build(BuildContext context) => SizedBox(
-        width: 520,
-        height: 326,
-        child: Stack(
-          fit: StackFit.expand,
-          children: [
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 72),
-              child: Column(
-                children: [
-                  const SizedBox(height: 40),
-                  Text(
-                    context.l10n.enterPinToDeleteAccount,
-                    style: TextStyle(
-                      color: context.theme.red,
-                      fontSize: 18,
-                    ),
-                  ),
-                  const SizedBox(height: 29),
-                  PinInputLayout(
-                    doVerify: (String pin) async {
-                      await context.accountServer.client.accountApi.deactive(
-                        DeactivateRequest(encryptPin(pin)!, verificationId),
-                      );
-                      Navigator.pop(context, true);
-                    },
-                  ),
-                  const SizedBox(height: 29),
-                  HighlightText(
-                    context.l10n.settingDeleteAccountPinContent(
-                      DateFormat.yMMMd().format(
-                        DateTime.now().add(const Duration(days: 30)),
-                      ),
-                    ),
-                    style: TextStyle(
-                      fontSize: 16,
-                      color: context.theme.text,
-                      height: 1.5,
-                    ),
-                    textAlign: TextAlign.center,
-                    highlightTextSpans: [
-                      HighlightTextSpan(
-                        context.l10n.learnMore,
-                        style: TextStyle(
-                          color: context.theme.accent,
-                        ),
-                        onTap: () => openUri(
-                            context, context.l10n.settingDeleteAccountUrl),
-                      ),
-                    ],
-                  ),
-                ],
-              ),
-            ),
-            const Align(
-              alignment: Alignment.topRight,
-              child: Padding(
-                padding: EdgeInsets.all(22),
-                child: MixinCloseButton(),
-              ),
-            ),
-          ],
-        ),
+  Widget build(BuildContext context) {
+    final textSpan = useMemoized(() {
+      final content = context.l10n.settingDeleteAccountPinContent(
+        DateFormat.yMMMd().format(DateTime.now().add(const Duration(days: 30))),
       );
+
+      final index = content.indexOf(context.l10n.learnMore);
+      if (index == -1) return TextSpan(text: content);
+
+      return TextSpan(
+        text: content.substring(0, index),
+        children: [
+          TextSpan(
+            text: context.l10n.learnMore,
+            style: TextStyle(
+              color: context.theme.accent,
+            ),
+            recognizer: TapGestureRecognizer()
+              ..onTap =
+                  () => openUri(context, context.l10n.settingDeleteAccountUrl),
+          ),
+          TextSpan(
+              text: content.substring(index + context.l10n.learnMore.length)),
+        ],
+      );
+    }, []);
+    return SizedBox(
+      width: 520,
+      height: 326,
+      child: Stack(
+        fit: StackFit.expand,
+        children: [
+          Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 72),
+            child: Column(
+              children: [
+                const SizedBox(height: 40),
+                Text(
+                  context.l10n.enterPinToDeleteAccount,
+                  style: TextStyle(
+                    color: context.theme.red,
+                    fontSize: 18,
+                  ),
+                ),
+                const SizedBox(height: 29),
+                PinInputLayout(
+                  doVerify: (String pin) async {
+                    await context.accountServer.client.accountApi.deactive(
+                      DeactivateRequest(encryptPin(pin)!, verificationId),
+                    );
+                    Navigator.pop(context, true);
+                  },
+                ),
+                const SizedBox(height: 29),
+                CustomText.rich(
+                  textSpan,
+                  style: TextStyle(
+                    fontSize: 16,
+                    color: context.theme.text,
+                    height: 1.5,
+                  ),
+                  textAlign: TextAlign.center,
+                ),
+              ],
+            ),
+          ),
+          const Align(
+            alignment: Alignment.topRight,
+            child: Padding(
+              padding: EdgeInsets.all(22),
+              child: MixinCloseButton(),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
 }
