@@ -1,4 +1,3 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_svg/svg.dart';
@@ -58,68 +57,65 @@ class MessagePost extends StatelessWidget {
   final bool clickable;
 
   @override
-  Widget build(BuildContext context) => SelectionArea(
-        selectionControls: _PostTextSelectionControls(),
-        child: InteractiveDecoratedBox(
-          onTap: clickable
-              ? () => PostPreview.push(context, message: context.message)
-              : null,
-          behavior: HitTestBehavior.deferToChild,
-          child: Container(
-            padding: padding,
-            decoration: decoration,
-            child: Stack(
-              children: [
-                HookBuilder(builder: (context) {
-                  final postContent =
-                      useMemoized(content.postOptimize, [content]);
+  Widget build(BuildContext context) => InteractiveDecoratedBox(
+        onTap: clickable
+            ? () => PostPreview.push(context, message: context.message)
+            : null,
+        behavior: HitTestBehavior.deferToChild,
+        child: Container(
+          padding: padding,
+          decoration: decoration,
+          child: Stack(
+            children: [
+              HookBuilder(builder: (context) {
+                final postContent =
+                    useMemoized(content.postOptimize, [content]);
 
-                  return ConstrainedBox(
-                    constraints: BoxConstraints(
-                      minHeight: showStatus ? 48 : 0,
-                      minWidth: 128,
-                      maxHeight: 400,
+                return ConstrainedBox(
+                  constraints: BoxConstraints(
+                    minHeight: showStatus ? 48 : 0,
+                    minWidth: 128,
+                    maxHeight: 400,
+                  ),
+                  child: ScrollConfiguration(
+                    behavior: ScrollConfiguration.of(context)
+                        .copyWith(scrollbars: false),
+                    child: SingleChildScrollView(
+                      physics: const NeverScrollableScrollPhysics(),
+                      child: MarkdownColumn(data: postContent),
                     ),
-                    child: ScrollConfiguration(
-                      behavior: ScrollConfiguration.of(context)
-                          .copyWith(scrollbars: false),
-                      child: SingleChildScrollView(
-                        physics: const NeverScrollableScrollPhysics(),
-                        child: MarkdownColumn(data: postContent),
-                      ),
-                    ),
-                  );
-                }),
+                  ),
+                );
+              }),
+              Positioned(
+                right: 0,
+                top: 0,
+                child: Container(
+                  decoration: _decoration,
+                  alignment: Alignment.center,
+                  child: SvgPicture.asset(
+                    Resources.assetsImagesPostDetailSvg,
+                    width: 20,
+                    height: 20,
+                  ),
+                ),
+              ),
+              if (showStatus)
                 Positioned(
                   right: 0,
-                  top: 0,
+                  bottom: 0,
                   child: Container(
                     decoration: _decoration,
-                    alignment: Alignment.center,
-                    child: SvgPicture.asset(
-                      Resources.assetsImagesPostDetailSvg,
-                      width: 20,
-                      height: 20,
+                    padding: const EdgeInsets.symmetric(
+                      horizontal: 6,
+                      vertical: 2,
+                    ),
+                    child: const MessageDatetimeAndStatus(
+                      color: Color.fromRGBO(255, 255, 255, 1),
                     ),
                   ),
                 ),
-                if (showStatus)
-                  Positioned(
-                    right: 0,
-                    bottom: 0,
-                    child: Container(
-                      decoration: _decoration,
-                      padding: const EdgeInsets.symmetric(
-                        horizontal: 6,
-                        vertical: 2,
-                      ),
-                      child: const MessageDatetimeAndStatus(
-                        color: Color.fromRGBO(255, 255, 255, 1),
-                      ),
-                    ),
-                  ),
-              ],
-            ),
+            ],
           ),
         ),
       );
@@ -181,31 +177,4 @@ class PostPreview extends StatelessWidget {
           ],
         ),
       );
-}
-
-class _PostTextSelectionControls extends TextSelectionControls {
-  @override
-  Widget buildHandle(BuildContext context, TextSelectionHandleType type,
-          double textLineHeight,
-          [VoidCallback? onTap]) =>
-      const SizedBox.shrink();
-
-  @override
-  Widget buildToolbar(
-          BuildContext context,
-          Rect globalEditableRegion,
-          double textLineHeight,
-          Offset position,
-          List<TextSelectionPoint> endpoints,
-          TextSelectionDelegate delegate,
-          ValueListenable<ClipboardStatus>? clipboardStatus,
-          Offset? lastSecondaryTapDownPosition) =>
-      const SizedBox.shrink();
-
-  @override
-  Offset getHandleAnchor(TextSelectionHandleType type, double textLineHeight) =>
-      Offset.zero;
-
-  @override
-  Size getHandleSize(double textLineHeight) => Size.zero;
 }
